updated some demo data
